SBERDOMA-62 extract INTEGRATION_CONTEXT_FIELD
