+++ apps/condo/domains/billing/schema/fields.js
@@ -1,0 +1,12 @@
+const INTEGRATION_CONTEXT_FIELD = {
+    schemaDoc: 'Integration context',
+    type: Relationship,
+    ref: 'BillingIntegrationOrganizationContext',
+    isRequired: true,
+    knexOptions: { isNotNullable: true }, // Required relationship only!
+    kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
+}
+
+module.exports = {
+    INTEGRATION_CONTEXT_FIELD,
+}
