+++ apps/condo/schema/Billing.js
@@ -3,6 +3,7 @@ const { Text, Relationship, Integer, DateTimeUtc, CalendarDay } = require('@keys
 const { GQLListSchema } = require('@core/keystone/schema')
 const { Json } = require('@core/keystone/fields')
 const { historical, versioned, uuided, tracked, softDeleted } = require('@core/keystone/plugins')
+const { INTEGRATION_CONTEXT_FIELD } = require('@condo/domains/billing/schema/fields')
 
 const { SENDER_FIELD, DV_FIELD, ORGANIZATION_OWNED_FIELD } = require('./_common')
 const { WRONG_TEXT_FORMAT, JSON_EXPECT_OBJECT_ERROR, DV_UNKNOWN_VERSION_ERROR, JSON_UNKNOWN_VERSION_ERROR } = require('../constants/errors')
@@ -10,15 +11,6 @@ const { UPPER_CASE_ALPHANUMERIC_REGEXP } = require('../constants/regexps')
 const { rules } = require('../access')
 const { hasRequestAndDbFields, hasValidJsonStructure } = require('../utils/validation.utils')
 
-
-const INTEGRATION_CONTEXT_FIELD = {
-    schemaDoc: 'Integration context',
-    type: Relationship,
-    ref: 'BillingIntegrationOrganizationContext',
-    isRequired: true,
-    knexOptions: { isNotNullable: true }, // Relationship only!
-    kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
-}
 
 const BILLING_PROPERTY_FIELD = {
     schemaDoc: 'Billing property',
